requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Session==0.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
PyMySQL==1.1.0
